#1 退货打印小票时报错闪退(待测) #2 商品数量允许输入小数

+++ app/src/main/java/com/ftrend/zgp/api/ShopCartContract.java
@@ -167,7 +167,7 @@ public interface ShopCartContract {
          *
          * @param num 购物车内的数量
          */
-        void updateTradeProd(double num, double price);
+        void updateTradeProd(long num, double price);
 
 
         /**
